https://grizzly.dev.java.net/issues/show_bug.cgi?id=483
selectionkeys used by comet in grizzly main selector are geting their attachment set to Long within a few ms after cometselector registration, this triggers cometselector idletimeout code path that detects Long attachment in the mainkey and cancels the cometkey and removes its attachment, leading to failed resumes due to cometselector.cancelkey returns false when the attachment already is null.

to mitigate this problem i set the attachment to "comet" instead of null in cometengine.handle
so is this a new problem, or has it existed all the way back to when i removed the bannedkey list in grizzly keep alive code ?. 
its very strange if this has managed to stay undetected for several months, it fails the comet unit tests rather badly.

later today i will be able to test on non linux platforms to see if the recent  cometcontext fixes are still needed or not. (cometD-client tests works rather well regardless of these fixes here, comet unit tests however dont)

git-svn-id: https://svn.java.net/svn/grizzly~svn/trunk/code@2661 04fdb32c-e91d-0410-84f5-98e489d73e21