https://grizzly.dev.java.net/issues/show_bug.cgi?id=425
lowering gc overhead by storing each key interest type in different LTQs, removing the need for selectionkeyOp wrapper for many usage cases.
this also lowers thread contention both between producers, and between producers and the consumer.

todo:
investigate if its worthwile to swap LTQ references with a local one to completely avoid thread contention between the producers and the consumer
the cost for doing so is cheap, just a reference swap. volatile for contended LTQ and normal reference for the private/local one.

git-svn-id: https://svn.java.net/svn/grizzly~svn/trunk/code@2663 04fdb32c-e91d-0410-84f5-98e489d73e21