https://grizzly.dev.java.net/issues/show_bug.cgi?id=555
spincounter is now reset when selectime is larger then thresholdvalue.
the spinrate is now calculated as spins per second instead of number of spins  over a non defined timeframe.
exceptions in register of channels in the new selector is now try catched to allow the process to complete.

this patch fixes 3 broken SSL / async unit tests in framework for me on jdk 7b56  and 6u14b05 on win64.
I measured spinrate up to 150K/sec.

git-svn-id: https://svn.java.net/svn/grizzly~svn/trunk/code@3057 04fdb32c-e91d-0410-84f5-98e489d73e21