https://grizzly.dev.java.net/issues/show_bug.cgi?id=547
Removed the selectionkey.cancel() we did right before the socket.close().
The cancel allowed for the selector thread to trigger too early, while the 
other thread was still doing its socket.close(), hence becoming stuck at the 
synchronized (stateLock) in the SocketChannelImpl.

Instead we now let the natural order of close logic do the cancel:
socket.close  closes the channel and the channel cancels the key in  
implCloseChannel().


git-svn-id: https://svn.java.net/svn/grizzly~svn/trunk/code@3059 04fdb32c-e91d-0410-84f5-98e489d73e21