Incremental fix for https://grizzly.dev.java.net/issues/show_bug.cgi?id=733 (" autoConfigure of reader threads breaks compatibility ")
As reported on the user list, upgrading to 1.9.18 breaks a lot of applications. We also measured some performance regression when leader/follower gets enabled in GlassFish (only!), but playing safe for the 1.9.18 release, turning it off. 

Better documented ControllerConfig, and add a new com.sun.grizzly.displayConfiguration that will output the state of the Controller at runtime.



git-svn-id: https://svn.java.net/svn/grizzly~svn/trunk/code@3620 04fdb32c-e91d-0410-84f5-98e489d73e21