Fixed an issue parsing single-line HTTP headers.
When processing the bytes of the header value, the ParsingState.checkpoint was only incremented if a shift was detected.  Because of this, the checkpoint was at the initial state at the time the bytes were to be pushed to the buffer, no push occurred resulting in a zero-length buffer.

(Thanks Alexey!)

git-svn-id: svn+ssh://svn.java.net/grizzly~svn/branches/2dot0/code@4344 04fdb32c-e91d-0410-84f5-98e489d73e21