Incremental fix for GF issue https://glassfish.dev.java.net/issues/show_bug.cgi?id=11891.
It may be the case that when processing a POST request that is suspended, the post body hasn't
been read.  So instead of having the do/while loop continue and try to parse the post body as a 
request line, exit the loop and allow the async thread to perform the expected cleanup/setup for
the next request.

This fix is currently considered incremental due to the fact that with this new code in place, there
is an intermitten log message coming from KeepAliveThreadAttachement:126.  Not sure if this
is a side effect of this fix or the symptom from another.

git-svn-id: https://svn.java.net/svn/grizzly~svn/trunk/code@4581 04fdb32c-e91d-0410-84f5-98e489d73e21