Update SelectorThread/FileCache lookup logic.  In the case the listener is bound to 0.0.0.0 or 0:0:0:0:0:0:0:0 the hash code will be 0 effectively associating the listener with the port only.
When this occurs, we won't be able to lookup an existing entity using the local address of the socket since it will have be associated with a an adapter address.   So if we can't find a match on the address + port combination, make another lookup attempt using the port only.

git-svn-id: https://svn.java.net/svn/grizzly~svn/trunk/code@4679 04fdb32c-e91d-0410-84f5-98e489d73e21