Fix for issue discovered while testing the v3 admin gui using port unification redirecting to ssl.
When attempting to find the protocol based on decrypted ssl data, we'd start a read loop without checking to see if the input bytebuffer already had data that could be checked.  This would mean we could sit for 5 seconds for data that wouldn't come before the input BB was actually decrypted and the loop ran again.


The logic will now check to see if any data is preset before the protocol check loop, and if so, it will set decrypt and set the buffer up for protocol finding.

git-svn-id: https://svn.java.net/svn/grizzly~svn/trunk/code@5415 04fdb32c-e91d-0410-84f5-98e489d73e21