Optimized maxPendingWrite handling.
Previously the total number of bytes that could be pending to be written was Integer.MAX_VALUE.  This was effectively saying that it was unlimited.  However, after doing some comparisons with 1.9 and 2.0 it was found that the space reserve/release was accounting for around 5-7% of the benchmark.

Since 1.9 doesn't have similar functionality, we decided to change Integer.MAX_VALUE to -1 to disable this feature.  If disabled, the reserve/release calls won't be made.

git-svn-id: svn+ssh://svn.java.net/grizzly~svn/branches/2dot0/code@5975 04fdb32c-e91d-0410-84f5-98e489d73e21