Problem discovered with chunked transfer encoding while working on async http client implementation.
In the AHC test case, jetty sends the entire response on one chunk:

<http header>
\r\n
5\r\n
content\r\n
0\r\n
\r\n

When the ChunkedTransferEncoding parses the content chunk, we're left with a remainder of \r\n0\r\n\r\n, however, the parsing stops at this point and the isLast() flag on the HttpContent is not set, so grizzly is left hanging thinking it will receive more data.

The fix is to check the remainder to see if it represents the end of the transfer.  If so, then parse the trailer out, set the approriate flags and continue.