Fix for http://java.net/jira/browse/GRIZZLY-993.
Update ByteBufferWrapper.get(ByteBuffer) to call through to get(ByteBuffer, int, int) if there isn't enough room in the destination buffer.
Fixed get(ByteBuffer, int, int) to deal with this case properly.

After these changes, it was finally possible to run the large message benchmark with the ByteBufferManager in place.