[2.3.x] [2.3.x] + implement + checkpoint for issue #1527 https://java.net/jira/browse/GRIZZLY-1527 "Implement client-side Connection pool"
Add max reconnect attempts.  Defaults to five.
If reconnect feature is disabled and the pool is unable to connect, async pollers are notified of connection failure.
If reconnect feature is enabled and the pool is unable to connect and if the max number of connect attempts is exceeded, then async pollers will be notified of connection failure.