+ fix issue #25578378
"Grizzly simulates blocking read over NIO channel using temporary selectors. We 
register a channel we want to read from on the selector with OP_READ interest 
and waiting for channel to become readable. The selector immediately returns 
value 1, meaning that there is one channel ready to be processed, the 
selectedKeys collection returns single key, which is the key we registered 
for our channel. But when we try to read from the channel - we read 0 bytes."

The work around, once the issue is encountered, terminates 
current selector and takes the next one from the pool.