Fix NPE when accessing a registered resource
When registering a resource, OSGiMainHandler.registerResourceHandler may pass a null servletContext to the OSGiResourceHandler constructor. This causes a NPE when accessing the resource, thrown by OSGiResourceHandler.service. The fix is to use the same solution as in OSGiMainHandler.registerFilter, which first checks if the servletContext obtained from the mapper is null. If so, use the mapper to create a new servletContext based on the HttpContext and add it to the mapper.